- Fjernet kode relatert til "Åpne fil" funksjonalitet - Ryddet opp i prosjekt-struktur / endret navn på "ting" - util funksjoner samlet på et sted
