--- src/api/domain/LogResponse.ts
@@ -1,15 +1,0 @@
-export type LogResponse = {
-    id: string;
-    correlation_id: string;
-    environment: string;
-    person?: string;
-    date: string;
-    metadata: string
-    xml: string;
-    type: string;
-    transaction_id: string;
-    method: string;
-    from: string;
-    persons: string;
-
-}
