+++ src/api/domain/types/index.ts
@@ -98,3 +98,19 @@ export interface GuiModel {
     metadata: Metadata;
     fromFile?: boolean;
 }
+
+export type LogResponse = {
+    id: string;
+    correlation_id: string;
+    environment: string;
+    person?: string;
+    date: string;
+    metadata: string
+    xml: string;
+    type: string;
+    transaction_id: string;
+    method: string;
+    from: string;
+    persons: string;
+
+}
